02_meme_gen_starter/src/requirements.txt
pandas==0.23.4
Pillow==7.0.0
pycodestyle==2.5.0
Flask==1.1.1
python-docx==0.8.10